Add files via upload
This is a Yara Rule Scanner to make sure your yara rule works!

Make sure to have a yara.yar file in the same directory your executing the .py from in order for it to actually work.

Also make sure the executable your detecting with your yara rules is in the same directory your putting into the console.

Make sure to  also not scan a lot of files. Your downloads or a single folder should be enough.